fix: resolve compilation errors across codebase
- Skill.kt: Added matches() method to Skill data class for AdaptiveRouter compatibility
- DefaultAgentService.kt: Fixed taskId parameter passing to runAgentLoop, fixed procedureStepMatch logic, changed lastScreenHash to Long, fixed EngineHolder.canAcquireFast() call, fixed StuckDetector record call with Long parameter
- PreActionJudge.kt: Fixed judgeWithTimeout response handling with null-safe text access
- ClarificationManager.kt: Fixed ticker and remainingMs StateFlow implementation
- ServerTokenStore.kt: Fixed verifyToken logic with proper null check
- EngineHolder.kt: Added import for LocalEngineLease
- InteractionWatchdog.kt: Changed screenHash parameter from Int to Long for consistency

+++ app/src/main/java/com/returngift/agent/agent/loop/InteractionWatchdog.kt
@@ -73,7 +73,7 @@ class InteractionWatchdog(
         toolName: String,
         argsFingerprint: String,
         verification: ActionVerifier.VerificationResult,
-        screenHash: Int,
+        screenHash: Long,
         expectedForeground: String? = null,
         overlayPresent: Boolean = false
     ): Recovery {
